[Dropdown] - cleanups
